+++ packages/connectors/src/coinbaseWallet.ts
@@ -1,7 +1,7 @@
 import type {
-  CoinbaseWalletSDK,
   Preference,
   ProviderInterface,
+  createCoinbaseWalletSDK,
 } from '@coinbase/wallet-sdk'
 import {
   ChainNotConfiguredError,
@@ -15,6 +15,7 @@ import type {
 } from 'cbw-sdk'
 import {
   type AddEthereumChainParameter,
+  type Address,
   type Hex,
   type ProviderRpcError,
   SwitchChainError,
@@ -62,14 +63,16 @@ export function coinbaseWallet<version extends Version>(
 
 type Version4Parameters = Mutable<
   Omit<
-    ConstructorParameters<typeof CoinbaseWalletSDK>[0],
-    'appChainIds' // set via wagmi config
+    Parameters<typeof createCoinbaseWalletSDK>[0],
+    | 'appChainIds' // set via wagmi config
+    | 'preference'
   > & {
+    // TODO(v3): Remove `Preference['options']`
     /**
      * Preference for the type of wallet to display.
      * @default 'all'
      */
-    preference?: Preference['options'] | undefined
+    preference?: Preference['options'] | Compute<Preference> | undefined
   }
 >
 
@@ -78,26 +81,38 @@ function version4(parameters: Version4Parameters) {
     // for backwards compatibility
     close?(): void
   }
+  type Properties = {
+    connect(parameters?: {
+      chainId?: number | undefined
+      instantOnboarding?: boolean | undefined
+      isReconnecting?: boolean | undefined
+    }): Promise<{
+      accounts: readonly Address[]
+      chainId: number
+    }>
+  }
 
-  let sdk: CoinbaseWalletSDK | undefined
   let walletProvider: Provider | undefined
 
   let accountsChanged: Connector['onAccountsChanged'] | undefined
   let chainChanged: Connector['onChainChanged'] | undefined
   let disconnect: Connector['onDisconnect'] | undefined
 
-  return createConnector<Provider>((config) => ({
+  return createConnector<Provider, Properties>((config) => ({
     id: 'coinbaseWalletSDK',
     name: 'Coinbase Wallet',
     rdns: 'com.coinbase.wallet',
-    supportsSimulation: true,
     type: coinbaseWallet.type,
-    async connect({ chainId } = {}) {
+    async connect({ chainId, ...rest } = {}) {
       try {
         const provider = await this.getProvider()
         const accounts = (
           (await provider.request({
             method: 'eth_requestAccounts',
+            params:
+              'instantOnboarding' in rest && rest.instantOnboarding
+                ? [{ onboarding: 'instant' }]
+                : [],
           })) as string[]
         ).map((x) => getAddress(x))
 
@@ -171,27 +186,23 @@ function version4(parameters: Version4Parameters) {
     },
     async getProvider() {
       if (!walletProvider) {
-        // Unwrapping import for Vite compatibility.
-        // See: https://github.com/vitejs/vite/issues/9703
-        const CoinbaseWalletSDK = await (async () => {
-          const SDK = await import('@coinbase/wallet-sdk')
-          if (
-            typeof SDK.CoinbaseWalletSDK !== 'function' &&
-            typeof SDK.default === 'function'
-          )
-            return SDK.default
-          return SDK.CoinbaseWalletSDK
+        const preference = (() => {
+          if (typeof parameters.preference === 'string')
+            return { options: parameters.preference }
+          return {
+            ...parameters.preference,
+            options: parameters.preference?.options ?? 'all',
+          }
         })()
 
-        sdk = new CoinbaseWalletSDK({
+        const { createCoinbaseWalletSDK } = await import('@coinbase/wallet-sdk')
+        const sdk = createCoinbaseWalletSDK({
           ...parameters,
           appChainIds: config.chains.map((x) => x.id),
+          preference,
         })
 
-        walletProvider = sdk.makeWeb3Provider({
-          ...parameters,
-          options: parameters.preference ?? 'all',
-        })
+        walletProvider = sdk.getProvider()
       }
 
       return walletProvider
@@ -327,7 +338,6 @@ function version3(parameters: Version3Parameters) {
   return createConnector<Provider>((config) => ({
     id: 'coinbaseWalletSDK',
     name: 'Coinbase Wallet',
-    supportsSimulation: true,
     type: coinbaseWallet.type,
     async connect({ chainId } = {}) {
       try {