+++ packages/create-wagmi/src/cli.test.ts
@@ -1,7 +1,10 @@
+import {
+  type ExecSyncOptionsWithStringEncoding,
+  execSync,
+} from 'node:child_process'
+import { mkdirSync, readdirSync, writeFileSync } from 'node:fs'
+import { rm } from 'node:fs/promises'
 import { join } from 'node:path'
-import type { ExecaSyncReturnValue, SyncOptions } from 'execa'
-import { execaCommandSync } from 'execa'
-import fs from 'fs-extra'
 import pc from 'picocolors'
 import { afterEach, beforeAll, expect, test } from 'vitest'
 
@@ -12,45 +15,50 @@ const cliPath = join(__dirname, '../dist/esm/cli.js')
 const projectName = 'test-app'
 const genPath = join(__dirname, projectName)
 
-function run(args: string[], options: SyncOptions = {}): ExecaSyncReturnValue {
-  return execaCommandSync(`node ${cliPath} ${args.join(' ')}`, options)
+function run(
+  args: string[],
+  options: ExecSyncOptionsWithStringEncoding = { encoding: 'utf8' },
+): string {
+  return execSync(`node ${cliPath} ${args.join(' ')}`, options)
 }
 
 function createNonEmptyDir() {
-  fs.mkdirpSync(genPath)
+  mkdirSync(genPath, { recursive: true })
   const file = join(genPath, 'foo')
-  fs.writeFileSync(file, 'bar')
+  writeFileSync(file, 'bar')
 }
 
-beforeAll(() => {
-  execaCommandSync('pnpm --filter create-wagmi build')
-  fs.remove(genPath)
+beforeAll(async () => {
+  execSync('pnpm --filter create-wagmi build')
+  await rm(genPath, { recursive: true, force: true })
 })
-afterEach(() => fs.remove(genPath))
+afterEach(async () => {
+  await rm(genPath, { recursive: true, force: true })
+})
 
 test('prompts for the project name if none supplied', () => {
-  const { stdout } = run([])
+  const stdout = run([])
   expect(stdout).toContain('Project name:')
 })
 
 test('prompts for the framework if none supplied when target dir is current directory', () => {
-  fs.mkdirpSync(genPath)
-  const { stdout } = run(['.'], { cwd: genPath })
+  mkdirSync(genPath)
+  const stdout = run(['.'], { cwd: genPath, encoding: 'utf8' })
   expect(stdout).toContain('Select a framework:')
 })
 
 test('prompts for the framework if none supplied', () => {
-  const { stdout } = run([projectName])
+  const stdout = run([projectName])
   expect(stdout).toContain('Select a framework:')
 })
 
 test('prompts for the framework on not supplying a value for --template', () => {
-  const { stdout } = run([projectName, '--template'])
+  const stdout = run([projectName, '--template'])
   expect(stdout).toContain('Select a framework:')
 })
 
 test('prompts for the framework on supplying an invalid template', () => {
-  const { stdout } = run([projectName, '--template', 'unknown'])
+  const stdout = run([projectName, '--template', 'unknown'])
   expect(stdout).toContain(
     `"unknown" isn't a valid template. Please choose from below:`,
   )
@@ -58,18 +66,19 @@ test('prompts for the framework on supplying an invalid template', () => {
 
 test('asks to overwrite non-empty target directory', () => {
   createNonEmptyDir()
-  const { stdout } = run([projectName], { cwd: __dirname })
+  const stdout = run([projectName], { cwd: __dirname, encoding: 'utf8' })
   expect(stdout).toContain(`Target directory "${projectName}" is not empty.`)
 })
 
 test('asks to overwrite non-empty current directory', () => {
   createNonEmptyDir()
-  const { stdout } = run(['.'], { cwd: genPath })
+  const stdout = run(['.'], { cwd: genPath, encoding: 'utf8' })
   expect(stdout).toContain('Current directory is not empty.')
 })
 
-const templateFiles = fs
-  .readdirSync(join(cliPath, '../../../templates/vite-react'))
+const templateFiles = readdirSync(
+  join(cliPath, '../../../templates/vite-react'),
+)
   .map((filePath) => {
     if (filePath === '_gitignore') return '.gitignore'
     if (filePath === '_env.local') return '.env.local'
@@ -79,38 +88,41 @@ const templateFiles = fs
   .sort()
 
 test('successfully scaffolds a project based on vite-react starter template', () => {
-  fs.ensureDirSync(genPath)
-  const { stdout } = run([projectName, '--template', 'vite-react'], {
+  mkdirSync(genPath, { recursive: true })
+  const stdout = run([projectName, '--template', 'vite-react'], {
     cwd: __dirname,
+    encoding: 'utf8',
   })
-  const generatedFiles = fs.readdirSync(genPath).sort()
+  const generatedFiles = readdirSync(genPath).sort()
 
   expect(stdout).toContain(`Scaffolding project in ${genPath}`)
   expect(templateFiles).toEqual(generatedFiles)
 })
 
 test('works with the -t alias', () => {
-  fs.ensureDirSync(genPath)
-  const { stdout } = run([projectName, '-t', 'vite-react'], {
+  mkdirSync(genPath, { recursive: true })
+  const stdout = run([projectName, '-t', 'vite-react'], {
     cwd: __dirname,
+    encoding: 'utf8',
   })
-  const generatedFiles = fs.readdirSync(genPath).sort()
+  const generatedFiles = readdirSync(genPath).sort()
 
   expect(stdout).toContain(`Scaffolding project in ${genPath}`)
   expect(templateFiles).toEqual(generatedFiles)
 })
 
 test('uses different package manager', () => {
-  fs.ensureDirSync(genPath)
-  const { stdout } = run([projectName, '--bun', '-t', 'vite-react'], {
+  mkdirSync(genPath, { recursive: true })
+  const stdout = run([projectName, '--bun', '-t', 'vite-react'], {
     cwd: __dirname,
+    encoding: 'utf8',
   })
 
   expect(stdout).toContain('bun install')
 })
 
 test('shows help', () => {
-  const { stdout } = run(['--help'])
+  const stdout = run(['--help'])
   expect(
     stdout
       .replace(version, 'x.y.z')
@@ -128,11 +140,12 @@ test('shows help', () => {
       --pnpm                 Use pnpm as your package manager 
       --yarn                 Use yarn as your package manager 
       -h, --help             Display this message 
-      -v, --version          Display version number "
+      -v, --version          Display version number 
+    "
   `)
 })
 
 test('shows version', () => {
-  const { stdout } = run(['--version'])
+  const stdout = run(['--version'])
   expect(stdout).toContain(`create-wagmi/${version} `)
 })