+++ playgrounds/vite-vue/src/wagmi.ts
@@ -1,6 +1,6 @@
 import { http, createConfig, createStorage } from '@wagmi/vue'
 import { mainnet, optimism, sepolia } from '@wagmi/vue/chains'
-import { coinbaseWallet, walletConnect } from '@wagmi/vue/connectors'
+import { coinbaseWallet, metaMask, walletConnect } from '@wagmi/vue/connectors'
 
 export const config = createConfig({
   chains: [mainnet, sepolia, optimism],
@@ -9,6 +9,7 @@ export const config = createConfig({
       projectId: import.meta.env.VITE_WC_PROJECT_ID,
     }),
     coinbaseWallet({ appName: 'Vite Vue Playground', darkMode: true }),
+    metaMask(),
   ],
   storage: createStorage({ storage: localStorage, key: 'vite-vue' }),
   transports: {